+++ src/edu/ku/brc/specify/tests/DataBuilder.java
@@ -1,5 +1,6 @@
 package edu.ku.brc.specify.tests;
 
+import java.math.BigDecimal;
 import java.text.DateFormat;
 import java.util.Calendar;
 import java.util.Date;
@@ -24,7 +25,6 @@ import edu.ku.brc.specify.datamodel.Borrow;
 import edu.ku.brc.specify.datamodel.BorrowAgents;
 import edu.ku.brc.specify.datamodel.BorrowMaterial;
 import edu.ku.brc.specify.datamodel.BorrowReturnMaterial;
-//import edu.ku.brc.specify.datamodel.BorrowShipments;
 import edu.ku.brc.specify.datamodel.CatalogSeries;
 import edu.ku.brc.specify.datamodel.CollectingEvent;
 import edu.ku.brc.specify.datamodel.CollectingEventAttr;
@@ -1711,10 +1711,10 @@ public class DataBuilder
                                           final Double elevationAccuracy,
                                           final Integer originalLatLongUnit,
                                           final String latLongType,
-                                          final Double latitude1,
-                                          final Double longitude1,
-                                          final Double latitude2,
-                                          final Double longitude2,
+                                          final BigDecimal latitude1,
+                                          final BigDecimal longitude1,
+                                          final BigDecimal latitude2,
+                                          final BigDecimal longitude2,
                                           final String latLongMethod,
                                           final Double latLongAccuracy,
                                           final String datum,
